fix: organised namespaces properly in ManageNemsSubscription function, removed unused usings
