+++ application/CohortManager/src/Functions/DemographicServices/ManageNemsSubscription/HealthCheckFunction.cs
@@ -1,6 +1,5 @@
-namespace NHS.CohortManager.NEMSSubscriptionFunction;
+namespace NHS.CohortManager.DemographicServices.ManageNemsSubscription;
 
-using System.Net;
 using System.Threading.Tasks;
 using HealthChecks.Extensions;
 using Microsoft.Azure.Functions.Worker;
