+++ application/CohortManager/src/Functions/DemographicServices/ManageNemsSubscription/NemsSubscriptionManager.cs
@@ -1,20 +1,17 @@
-namespace NHS.CohortManager.DemographicServices;
+namespace NHS.CohortManager.DemographicServices.ManageNemsSubscription;
 
 using System;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Collections.Generic;
-using Azure;
-using Azure.Data.Tables;
 using Microsoft.Extensions.Logging;
-using System.Net.Http;
-using DataServices.Client;
 using Microsoft.Extensions.Options;
 using Model;
 using Common;
 using DataServices.Core;
 using System.Security.Cryptography.X509Certificates;
 using Hl7.Fhir.Serialization;
+using NHS.CohortManager.DemographicServices.ManageNemsSubscription.Config;
 // Use explicit STU3 aliases to avoid R4 conflicts
 using STU3Subscription = Hl7.Fhir.Model.Subscription;
 using STU3Meta = Hl7.Fhir.Model.Meta;
