+++ application/CohortManager/src/Functions/DemographicServices/ManageNemsSubscription/Program.cs
@@ -3,9 +3,9 @@ using HealthChecks.Extensions;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Common;
-using NHS.CohortManager.DemographicServices;
+using NHS.CohortManager.DemographicServices.ManageNemsSubscription.Config;
+using NHS.CohortManager.DemographicServices.ManageNemsSubscription;
 using DataServices.Database;
-using Model;
 using Microsoft.Extensions.Logging;
 using System.Security.Cryptography.X509Certificates;
 using Azure.Security.KeyVault.Certificates;
