+++ tests/UnitTests/DemographicServicesTests/ManageNemsSubscriptionTests/ManageNemsSubscriptionTests.cs
@@ -7,7 +7,8 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Model;
 using Moq;
-using NHS.CohortManager.DemographicServices;
+using NHS.CohortManager.DemographicServices.ManageNemsSubscription;
+using NHS.CohortManager.DemographicServices.ManageNemsSubscription.Config;
 using DataServices.Core;
 using System.Linq.Expressions;
 using System.Collections.Specialized;
